requirements.txt
google-generativeai
duckduckgo-search
requests
pyyaml
markdown
